requirements.txt
torch
pandas
scikit-learn
transformers==4.57.3
pillow==11.3.0
faiss-cpu==1.13.2
gradio==6.10.0